kafka_setup/requirements.txt
kafka-python
pydantic
faker
